Added float support

+++ report/item.go
@@ -5,6 +5,6 @@ import "time"
 // Item is a stuct holding information about the time worked on a task on a given day.
 type Item struct {
 	Date  time.Time
-	Hours int
+	Hours float64
 	Task  string
 }
